+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to Rainbow-Node-SDK will be documented in this file.
 
 ## [1.86.0-dotnet.0] - 2021-03-XX
 -   Add API `ConversationsService::resetHistoryPageForConversation` to restart from scratch the retrieve of messages from history with getHistoryPage.
+-   Fix the resources information when getContactByJid or by Id for the connected user.
 
 ## [1.85.1] - 2021-03-18
 -   Delivery 1.85.0 failed, so new one.